fix(ios): cfm 崩溃——全链 dtype 统一，Ops 输出还原输入 dtype + 拼接点显式对齐
根因：lengthRegulate 输出为权重 dtype(F16)，与 F32 zeros/条件拼接时 MLX 不提升 dtype → fatal。
修复：Ops.linear/conv1d 计算用权重 dtype、输出还原输入 dtype；TTSPipeline/cfm/DiT 拼接点显式 asType 对齐。

+++ ios/Sources/MLXIndexTTS2Core/TTSPipeline.swift
@@ -208,11 +208,11 @@ public final class TTSPipeline {
             let catMu: MLXArray
             let promptMel: MLXArray
             if let pb = prompt {
-                catMu = MLX.concatenated([pb.promptCondition, cond], axis: 1)
+                catMu = MLX.concatenated([pb.promptCondition.asType(cond.dtype), cond], axis: 1)
                 promptMel = pb.refMel
             } else {
                 // 零占位 prompt（管线验证路径）
-                let zeros = MLXArray.zeros([1, promptLen, 512])
+                let zeros = MLXArray.zeros([1, promptLen, 512]).asType(cond.dtype)   // ⚠️ 与 cond 同 dtype
                 catMu = MLX.concatenated([zeros, cond], axis: 1)
                 promptMel = MLXArray.zeros([1, 80, promptLen])
             }
